Extract HtmlSanitizer from StructuredPopup
Move the HTML-sanitization logic out of StructuredPopup (whose job is popup
structure) into a dedicated, separately-tested Maps\Presentation\HtmlSanitizer
that StructuredPopup delegates to. Behavior-preserving: the sanitization tests
move to HtmlSanitizerTest, leaving StructuredPopupTest focused on structure
plus checks that property values, names, and the title are sanitized.
