+++ src/Map/StructuredPopup.php
@@ -4,15 +4,13 @@ declare( strict_types = 1 );
 
 namespace Maps\Map;
 
-use DOMDocument;
-use DOMElement;
+use Maps\Presentation\HtmlSanitizer;
 
 class StructuredPopup {
 
-	private const SAFE_URL_SCHEMES = [ 'http', 'https', 'ftp', 'ftps', 'mailto', 'tel' ];
-
 	private string $titleHtml;
 	private array $propertyValues;
+	private HtmlSanitizer $sanitizer;
 
 	/**
 	 * @param string[] $propertyValues
@@ -20,10 +18,11 @@ class StructuredPopup {
 	public function __construct( string $titleHtml, array $propertyValues ) {
 		$this->titleHtml = $titleHtml;
 		$this->propertyValues = $propertyValues;
+		$this->sanitizer = new HtmlSanitizer();
 	}
 
 	public function getHtml(): string {
-		$title = $this->sanitize( $this->titleHtml );
+		$title = $this->sanitizer->sanitize( $this->titleHtml );
 		$valueList = $this->getPropertyValueList();
 		$separator = $title === '' || $valueList === '' ? '' : '<br>';
 
@@ -34,7 +33,7 @@ class StructuredPopup {
 		$lines = [];
 
 		foreach ( $this->propertyValues as $name => $value ) {
-			$lines[] = $this->bold( $this->sanitize( (string)$name ) ) . ': ' . $this->sanitize( $value );
+			$lines[] = $this->bold( $this->sanitizer->sanitize( (string)$name ) ) . ': ' . $this->sanitizer->sanitize( $value );
 		}
 
 		return implode( '<br>', $lines );
@@ -42,75 +41,6 @@ class StructuredPopup {
 
 	private function bold( string $html ): string {
 		return '<strong>' . $html . '</strong>';
-	}
-
-	/**
-	 * Keeps only <a> and <img> tags and strips event-handler attributes and unsafe URL schemes
-	 * from them. Relative and http(s) URLs are preserved so links to wiki pages and images that
-	 * the popup is meant to show keep working.
-	 */
-	private function sanitize( string $html ): string {
-		$html = strip_tags( $html, '<a><img>' );
-
-		if ( strpos( $html, '<' ) === false ) {
-			return $html;
-		}
-
-		return $this->stripDangerousAttributes( $html );
-	}
-
-	private function stripDangerousAttributes( string $html ): string {
-		$document = new DOMDocument();
-
-		$previousErrorHandling = libxml_use_internal_errors( true );
-		$loaded = $document->loadHTML(
-			'<?xml encoding="UTF-8"?><div>' . $html . '</div>',
-			LIBXML_HTML_NOIMPLIED | LIBXML_HTML_NODEFDTD
-		);
-		libxml_clear_errors();
-		libxml_use_internal_errors( $previousErrorHandling );
-
-		if ( !$loaded || $document->documentElement === null ) {
-			return strip_tags( $html );
-		}
-
-		foreach ( $document->getElementsByTagName( '*' ) as $element ) {
-			$this->stripDangerousAttributesFromElement( $element );
-		}
-
-		return $this->innerHtml( $document->documentElement );
-	}
-
-	private function stripDangerousAttributesFromElement( DOMElement $element ): void {
-		foreach ( iterator_to_array( $element->attributes ) as $attribute ) {
-			$name = strtolower( $attribute->name );
-
-			if ( strpos( $name, 'on' ) === 0 ) {
-				$element->removeAttribute( $attribute->name );
-			} elseif ( ( $name === 'href' || $name === 'src' ) && $this->hasUnsafeScheme( $attribute->value ) ) {
-				$element->removeAttribute( $attribute->name );
-			}
-		}
-	}
-
-	private function hasUnsafeScheme( string $url ): bool {
-		$normalized = strtolower( (string)preg_replace( '/[\x00-\x20]+/', '', $url ) );
-
-		if ( preg_match( '/^[a-z][a-z0-9+.\-]*:/', $normalized ) !== 1 ) {
-			return false;
-		}
-
-		return !in_array( strstr( $normalized, ':', true ), self::SAFE_URL_SCHEMES, true );
-	}
-
-	private function innerHtml( DOMElement $element ): string {
-		$html = '';
-
-		foreach ( $element->childNodes as $child ) {
-			$html .= $element->ownerDocument->saveHTML( $child );
-		}
-
-		return $html;
 	}
 
 }
